requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai
pydantic
pydantic_settings
langchain_groq 
langchain_community 
faiss-cpu 
pymongo 
langchain-mongodb
huggingface_hub
sentence_transformers
langchain_google_genai
langchain_huggingface